// web/remote-control/src/generated/types/ClientEnvelope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClientMessage } from "./ClientMessage";
import type { RequestId } from "./RequestId";

export type ClientEnvelope = { version: number, request_id: RequestId, payload: ClientMessage, };