// web/remote-control/src/generated/types/ServerEnvelope.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RequestId } from "./RequestId";
import type { ServerMessage } from "./ServerMessage";

export type ServerEnvelope = { version: number, request_id: RequestId | null, sequence: number, payload: ServerMessage, };